Revert "Feat: Player 생성 전 SequencePlayer 분리"
This reverts commit f444b8c95676cbca9ee7ed52e994eb373d05b36b.
