+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/GameFramework/Character.h
@@ -1,9 +1,11 @@
 ﻿#pragma once
 #include "Pawn.h"
 
+
 class ACharacter : public APawn
 {
     DECLARE_CLASS(ACharacter, APawn)
+
 public:
     ACharacter() = default;
