+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/World/World.cpp
@@ -3,7 +3,6 @@
 #include "CollisionManager.h"
 #include "PhysicsManager.h"
 #include "Actors/Player.h"
-#include "GameFramework/Character.h"
 #include "BaseGizmos/TransformGizmo.h"
 #include "Classes/Components/StaticMeshComponent.h"
 #include "Engine/FObjLoader.h"
@@ -183,7 +182,7 @@ UWorld* UWorld::GetWorld() const
     return const_cast<UWorld*>(this);
 }
 
-ACharacter* UWorld::GetMainPlayer() const
+APlayer* UWorld::GetMainPlayer() const
 {
     if (MainPlayer)
     {
@@ -191,7 +190,7 @@ ACharacter* UWorld::GetMainPlayer() const
     }
     
     //메인플레이어 설정안하면 있는거중 한개
-    for (const auto Iter: TObjectRange<ACharacter>())
+    for (const auto Iter: TObjectRange<APlayer>())
     {
         if (Iter->GetWorld() == GEngine->ActiveWorld)
         {
