+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/World/World.h
@@ -10,7 +10,6 @@
 #include "Engine/EventManager.h"
 #include "UObject/UObjectIterator.h"
 
-class ACharacter;
 class UPrimitiveComponent;
 struct FOverlapResult;
 class UCameraComponent;
@@ -70,8 +69,8 @@ public:
     
     FEventManager EventManager;
 
-    void SetMainPlayer(ACharacter* InPlayer){ MainPlayer = InPlayer; }
-    ACharacter* GetMainPlayer() const;
+    void SetMainPlayer(APlayer* InPlayer){ MainPlayer = InPlayer; }
+    APlayer* GetMainPlayer() const;
 
     void SetPlayerController(APlayerController* InPlayerController){ PlayerController = InPlayerController; }
     APlayerController* GetPlayerController() const;
@@ -99,7 +98,7 @@ private:
     // TODO: 싱글 플레이어면 상관 없지만, 로컬 멀티 플레이어인 경우를 위해 배열로 관리하는 방법을 고려하기.
     APlayerController* PlayerController = nullptr;
 
-    ACharacter* MainPlayer = nullptr;
+    APlayer* MainPlayer = nullptr;
 
     UTextComponent* MainTextComponent = nullptr;
